+++ CHANGELOG.md
@@ -6,6 +6,14 @@ All notable changes to the "**Evilbot**" repository will be documented in this f
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+---
+## [**4.3.2**] - 2024-03-01
+
+### Fixed
+* Fixed voice channel checks for music module
+* Fixed a bug where the settings were not loaded when starting the bot, but only when any parameter was updated for the first time
+
+
 ## [**4.3.1**] - 2024-02-29
 
 ### Added